rename templates to prompts in ai crate
