+++ crates/ai/src/ai.rs
@@ -1,4 +1,4 @@
 pub mod completion;
 pub mod embedding;
 pub mod models;
-pub mod templates;
+pub mod prompts;
