+++ crates/ai/src/prompts/mod.rs
